requirements.txt
# requirements.txt
Django==4.2.7
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
djangorestframework==3.14.0
python-decouple==3.8
stripe==7.8.0
django-cors-headers==4.3.1